.ipynb_checkpoints/requirements-checkpoint.txt
torch>=2.3.0
torchvision>=0.18.0
fastapi
uvicorn[standard]
streamlit
pillow
numpy
matplotlib
scikit-learn
seaborn
locust